fix: move commitments backup folder to utils
