+++ wallet/src/utils/CommitmentsBackup/commitmentsVerification.js
