+++ wallet/tests/unit/utils/CommitmentsBackup/commitmentsVerification.test.js
@@ -1,4 +1,4 @@
-import isCommitmentsCPKDMatchDerivedKeys from '../../../../src/useCases/CommitmentsBackup/commitmentsVerification';
+import isCommitmentsCPKDMatchDerivedKeys from '../../../../src/utils/CommitmentsBackup/commitmentsVerification';
 
 const wrongCommitments = [
   {
