+++ wallet/tests/unit/utils/CommitmentsBackup/export.test.js
@@ -1,7 +1,7 @@
 import indexedDB from 'fake-indexeddb';
 import 'fake-indexeddb/auto';
 
-import exportIndexdDB from '../../../../src/useCases/CommitmentsBackup/export';
+import exportIndexdDB from '../../../../src/utils/CommitmentsBackup/export';
 
 describe('Tests about indexedDB', () => {
   const OBJ_STORE_NAME = 'commitments';
